requirements.txt
langchain
python-decouple
langchain-google-genai
fastapi
uvicorn
